+++ feedsx/dependencies.gradle
@@ -80,5 +80,5 @@ dependencies {
     implementation feed_deps.cropper.main
 
     // feed-sdk
-    implementation 'community.likeminds:likemindsfeed:1.5.0'
+    implementation 'community.likeminds:likemindsfeed:1.6.0'
 }
