+++ feedsx/src/main/java/com/likeminds/feedsx/post/create/view/LMFeedCreatePostFragment.kt
@@ -343,7 +343,6 @@ class LMFeedCreatePostFragment :
     private fun initiateMediaPicker(list: List<String>) {
         val extras = MediaPickerExtras.Builder()
             .mediaTypes(list)
-            .allowMultipleSelect(true)
             .build()
         val intent = LMFeedMediaPickerActivity.getIntent(requireContext(), extras)
         galleryLauncher.launch(intent)
