+++ feedsx/src/main/java/com/likeminds/feedsx/post/detail/view/adapter/databinder/ItemPostDetailCommentViewDataBinder.kt
@@ -63,7 +63,7 @@ class ItemPostDetailCommentViewDataBinder constructor(
             tvCommenterName.text = data.user.name
 
             tvCommenterName.setOnClickListener {
-                LikeMindsFeedUI.lmFeedListener.openProfile(
+                SDKApplication.getLMFeedUICallback()?.openProfile(
                     data.user.sdkClientInfoViewData.uuid,
                     data.id,
                     LMFeedAnalytics.Source.FEED
